requirements.txt
azure-functions==1.18.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dill==0.3.8
Flask==3.0.2
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.1
pickle4==0.0.1
pyodbc==5.1.0
python-dateutil==2.9.0
pytz==2024.1
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
SQLAlchemy==2.0.28
threadpoolctl==3.3.0
typing_extensions==4.10.0
tzdata==2024.1
Werkzeug==3.0.1
xgboost==2.0.3
